requirements.txt
alabaster==0.7.12
Babel==2.9.1
bootstrap-py==1.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
defusedxml==0.7.1
docutils==0.17.1
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
packaging==21.3
pguard==0.4.0
protobuf==3.19.4
Pygments==2.11.2
PyMySQL==1.0.2
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.35
urllib3==1.26.9
Werkzeug==2.0.2
zipp==3.8.0
